BREAKING: feat: make lambda pick up the script from the request body
